Extract interface: ConigurationRepository
